Refactor SSOBlock
