+++ front/app/components/admin/ActionForm/FlowVisualization/Blocks/index.tsx
@@ -1,27 +1,20 @@
 import React, { useState } from 'react';
 
-import { Box, Text } from '@citizenlab/cl2-component-library';
-
 import { IPermissionsCustomFieldData } from 'api/permissions_custom_fields/types';
 import { PermittedBy } from 'api/phase_permissions/types';
-import useVerificationMethodVerifiedActions from 'api/verification_methods/useVerificationMethodVerifiedActions';
 
 import useFeatureFlag from 'hooks/useFeatureFlag';
-import useLocalize from 'hooks/useLocalize';
 
 import { useIntl } from 'utils/cl-intl';
 
 import { getNumberOfVerificationLockedItems } from '../../utils';
 
-import { Block, SSOBlock } from './Block';
+import Block from './Block';
 import Edge from './Edge';
 import messages from './messages';
+import SSOBlock from './SSOBlock';
 import SSOConfigModal from './SSOConfigModal';
-import {
-  getReturnedFieldsPreview,
-  enabledSteps,
-  getVerificationFrequencyExplanation,
-} from './utils';
+import { enabledSteps } from './utils';
 
 interface Props {
   permittedBy: PermittedBy;
@@ -38,12 +31,10 @@ const Blocks = ({
   verificationExpiry,
   onChangeVerificationExpiry,
 }: Props) => {
-  const { data: verificationMethod } = useVerificationMethodVerifiedActions();
   const emailConfirmationEnabled = useFeatureFlag({
     name: 'user_confirmation',
   });
   const { formatMessage } = useIntl();
-  const localize = useLocalize();
   const [modalOpen, setModalOpen] = useState(false);
 
   const numberOfVerificatiomLockedItems = getNumberOfVerificationLockedItems(
@@ -52,9 +43,6 @@ const Blocks = ({
 
   const showCustomFields =
     permissionsCustomFields.length > numberOfVerificatiomLockedItems;
-
-  const verificationMethodMetadata =
-    verificationMethod?.data.attributes.action_metadata;
 
   if (permittedBy === 'admins_moderators') return null;
 
@@ -137,46 +125,11 @@ const Blocks = ({
     );
   }
 
-  if (permittedBy === 'verified' && verificationMethodMetadata) {
-    const authenticateMessage = formatMessage(
-      messages.authenticateWithVerificationProvider,
-      {
-        verificationMethod: verificationMethodMetadata.name,
-      }
-    );
-
-    const verifiedFieldsMessage = formatMessage(messages.verifiedFields);
-
-    const verificationFrequencyExplanation =
-      getVerificationFrequencyExplanation(verificationExpiry, formatMessage);
-
-    const returnedFieldsPreview = getReturnedFieldsPreview(
-      verificationMethodMetadata,
-      localize
-    );
-
+  if (permittedBy === 'verified') {
     return (
       <>
         <SSOBlock
-          number={1}
-          text={
-            <>
-              {authenticateMessage}
-              <br />
-              {verificationFrequencyExplanation && (
-                <>
-                  <Text fontSize="xs" mt="8px" color="primary">
-                    {verificationFrequencyExplanation}
-                  </Text>
-                </>
-              )}
-              <Box mt="12px">
-                {verifiedFieldsMessage}
-                <br />
-                <b>{returnedFieldsPreview}</b>
-              </Box>
-            </>
-          }
+          verificationExpiry={verificationExpiry}
           onClick={() => setModalOpen(true)}
         />
         {showCustomFields && (
@@ -188,15 +141,12 @@ const Blocks = ({
             />
           </>
         )}
-        {verificationMethodMetadata && (
-          <SSOConfigModal
-            opened={modalOpen}
-            onClose={() => setModalOpen(false)}
-            verificationExpiry={verificationExpiry}
-            verificationMethodMetadata={verificationMethodMetadata}
-            onChangeVerificationExpiry={onChangeVerificationExpiry}
-          />
-        )}
+        <SSOConfigModal
+          opened={modalOpen}
+          onClose={() => setModalOpen(false)}
+          verificationExpiry={verificationExpiry}
+          onChangeVerificationExpiry={onChangeVerificationExpiry}
+        />
       </>
     );
   }
